+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.2
+
+* Add Swift Package Manager support while preserving CocoaPods support.
+* Update the minimum supported Flutter and Dart versions.
+
 ## 0.0.1
 
 * TODO: Describe initial release.